disable goffuzz in binaries
