+++ staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/BUILD
@@ -47,9 +47,11 @@ go_library(
         "labels.go",
         "meta.go",
         "micro_time.go",
+        "micro_time_fuzz.go",
         "micro_time_proto.go",
         "register.go",
         "time.go",
+        "time_fuzz.go",
         "time_proto.go",
         "types.go",
         "types_swagger_doc_generated.go",
