+++ staging/src/k8s.io/apimachinery/pkg/util/intstr/BUILD
@@ -17,6 +17,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "generated.pb.go",
+        "instr_fuzz.go",
         "intstr.go",
     ],
     importmap = "k8s.io/kubernetes/vendor/k8s.io/apimachinery/pkg/util/intstr",
